expense_tracker/api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
SQLAlchemy==2.0.23
itsdangerous==2.1.2
